feat: 출발 시간 계산 로직 구현

+++ backend/src/main/java/com/ody/route/domain/Duration.java
@@ -7,5 +7,5 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public class Duration {
 
-    private final int minutes;
+    private final long minutes;
 }
